Fixed a long standing bug in PImplHelper - since it was allowed to use derived types - That broke all copy & move function since only the base would be copied.
This is now hopefully solved at the cost of complexity - if the user wants to use derived types they must now specify that in an additional flag template parameter,
and if used additional information about the currently stored type will be stored (such as function pointers to copy/move constructors).
This change also brought the need to write a test for the implementation - I added a basic one, and if any future bugs appear  I add them to the test.