Added a implementation of IteratorAdopter to make it easier to create iterators.
Currently it supports implementing
  - forward_iterator_tag
  - bidirectional_iterator_tag
  - random_access_iterator_tag.
It is designed to have as simple of a interface as I could do, and to make a fully random access iterator, only 6 functions have to be implemented.
 * All iterators:
    - equal
    - increment
    - dereference
 * bidirectional
    - decrement
* random access
   - advance
   - distance

The usage is quite simple:
  class Iterator :
       public IteratorAdopter<Iterator, Type, Tag>
  {
    [...]
  };

It also supports iterators to 'temporary' objects (in those cases the 'dereference' function don't return a reference), in that case 'operator ->' will be disabled, and 'operator *' will return a temporary object.